made simple connection work on multiple clients

+++ LudoServer/Handlers/DeleteGameHandler.cs
@@ -1,4 +1,5 @@
 ﻿using LudoServer.Services;
+using LudoServer.Session;
 using SharedModels.Models;
 using SharedModels.Models.DTO;
 using SharedModels.TransferMsg;
@@ -6,6 +7,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
+using System.Text.Json;
 using System.Threading.Tasks;
 
 namespace LudoServer.Handlers
@@ -15,13 +17,15 @@ namespace LudoServer.Handlers
         private readonly GameService _gameService;
         private readonly PlayerService _playerService;
         private readonly GamePieceService _gamePieceService;
+        private readonly GameSessionManager _gameSessionManager;
         public string MessageType => "DeleteGame";
 
-        public DeleteGameHandler(GameService gameService, PlayerService playerService, GamePieceService gamePieceService)
+        public DeleteGameHandler(GameService gameService, PlayerService playerService, GamePieceService gamePieceService, GameSessionManager gameSessionManager)
         {
             _gameService = gameService;
             _playerService = playerService;
             _gamePieceService = gamePieceService;
+            _gameSessionManager = gameSessionManager;
         }
 
         public async Task<string> HandleAsync(string payload)
@@ -29,13 +33,18 @@ namespace LudoServer.Handlers
             var game = System.Text.Json.JsonSerializer.Deserialize<Game>(payload);
 
             if (game == null) return "game is null";
-
-
+            
+            var session = _gameSessionManager.GetSession(game);
+            if (session != null)
+            {
+                return JsonSerializer.Serialize("Game is active, cannot be deleted");
+                
+            }
             //call gameservices (Should be transaction, but w/e)
             _gamePieceService.deleteGamePiecesfromGame(game);
             _playerService.deletePlayersfromGame(game);
             _gameService.delete(game);
-            return "GameDeleted";
+            return JsonSerializer.Serialize("GameDeleted");
         }
